Basic prints done
It now prints out the basic time, but the different amounts (years, days, hours, etc), at the moment should not be added.
In the future the total should be split across the categories. So if you add them all back together, you get the total. RN you can not add them together.